asdfa

+++ src/components/WorkoutDetails.jsx
@@ -1,15 +1,19 @@
 import React from "react";
 import { useWorkoutContext } from "../hooks/useWorkoutContext";
-
+import { useAuthContext } from "../hooks/useAuthContext";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 const API_URL = process.env.REACT_APP_API_URL;
 export default function WorkoutDetails({ workout }) {
   const { dispatch } = useWorkoutContext();
-
+  const { user } = useAuthContext();
   const handleClick = async () => {
+    if (!user) return;
     const response = await fetch(`${API_URL}/api/workouts/` + workout._id, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     const json = await response.json();
     if (response.ok) {
@@ -30,7 +34,7 @@ export default function WorkoutDetails({ workout }) {
       <p>
         {formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}
       </p>
-      <span class="material-symbols-outlined" onClick={handleClick}>
+      <span className="material-symbols-outlined" onClick={handleClick}>
         delete
       </span>
     </div>
